requirements.txt
# Core Data Processing
pandas==2.1.4
numpy==1.26.2
scipy==1.11.4

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0

# Visualization
plotly==5.18.0
matplotlib==3.8.2
seaborn==0.13.0
streamlit==1.29.0
altair==5.2.0

# Machine Learning
scikit-learn==1.3.2
statsmodels==0.14.1

# Data Validation
great-expectations==0.18.8

# Web Scraping (for NSE data download)
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.16.0
lxml==4.9.3

# File Processing
openpyxl==3.1.2
python-docx==1.1.0
PyPDF2==3.0.1

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
python-dateutil==2.8.2
pytz==2023.3

# Workflow Orchestration
apache-airflow==2.8.0
prefect==2.14.13

# API Framework
fastapi==0.109.0
uvicorn==0.25.0
pydantic==2.5.3

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Code Quality
black==23.12.1
ruff==0.1.9
mypy==1.7.1

# Monitoring
prometheus-client==0.19.0

# Caching
redis==5.0.1

# Time Series
tslearn==0.6.2